requirements.txt
names~=0.3.0
curl_cffi==0.10.0
pytz~=2024.1
loguru~=0.7.2
httpx~=0.27.0
urllib3~=2.2.2
art~=6.2
PyYAML~=6.0.2
pydantic~=2.8.2
inquirer~=3.3.0
colorama~=0.4.6
better_proxy
tortoise-orm
imap_tools
numpy~=1.26.4
aiofiles~=24.1.0
aiocsv~=1.3.2
rich~=13.9.2
python_socks
asyncpg
openpyxl~=3.1.5
imap-tools~=1.10.0
python-socks~=2.7.1
requests~=2.32.3